fixed stuff, added chroma, attempted a map height finder thing (not finished, and the data I have is for 4v4s only)

+++ src/main/java/net/iridescent/wyvtilities/hud/elements/impl/ElementComboDisplay.java
@@ -44,7 +44,7 @@ public class ElementComboDisplay extends Elements {
         if (System.currentTimeMillis() - lastHit > 3000)
             combo = 0;
         this.setRenderedValue(combo == 0 ? "none" : combo + " Combo");
-        this.height = 10 * this.getPosition().getScale();
+        this.height = (int) (10 * this.getPosition().getScale());
         super.onRendered(position);
     }
